requirements.txt
numpy
opencv-python
matplotlib
pygame
PyOpenGL
PyOpenGL_accelerate
